temp save -- Compartment inheritance
git-tfs-id: [http://tfs.bumc.bu.edu:8080/tfs/BU-MICRO-MSI]$/Daphne/Daphne-axin2;C5946

+++ NativeDaphneLibrary/NtCollisionManager.h
@@ -66,7 +66,7 @@ namespace NativeDaphneLibrary
 
 		int getPairCount()
 		{
-			return pairs->size();
+			return (int)pairs->size();
 		}
 
 		bool itemExists(int key)
